# Backend/MainService/openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 1.0.0
  description: OpenAPI спецификация для основного сервиса
paths:
...
  /api/v1/main_service/projects/{projectId}/members:
    get:
      tags:
        - ProjectMember
      summary: Получение участников проекта
      operationId: getProjectMembers
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMemberFullResponse'
    post:
      tags:
        - ProjectMember
      summary: Добавление участника в проект
      operationId: addProjectMember
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMemberRequest'
      responses:
        '201':
          description: Участник добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberFullResponse'
  /api/v1/main_service/projects/{projectId}/members/{memberId}:
    delete:
      tags:
        - ProjectMember
      summary: Удаление участника из проекта
      operationId: removeProjectMember
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Участник удалён
          content: { }
...
  /api/v1/main_service/projects:
    post:
      tags:
        - Project
      summary: Создание проекта
      description: Создает новый проект. Использует имя пользователя из контекста безопасности.
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/XUserNameHeader'
      requestBody:
        $ref: '#/components/requestBodies/ProjectRequestWithImage'
      responses:
        '201':
          description: Проект успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFullResponse'
    get:
      tags:
        - Project
      summary: Получение списка проектов
      description: Возвращает список проектов с поддержкой фильтрации по категории, поисковому запросу и статусу модерации. Поддерживает постраничный вывод.
      operationId: getProjects
      parameters:
        - name: seekingRole
          in: query
          schema:
            $ref: '#/components/schemas/SeekingRole'
        - name: category
          in: query
          schema:
            type: string
        - name: domain
          in: query
          schema:
            type: string
        - name: query
          in: query
          description: Поисковый запрос по названию проекта
          schema:
            type: string
        - name: moderationPassed
          in: query
          description: Фильтр по статусу модерации
          schema:
            type: boolean
          required: true
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          description: Размер страницы
          required: true
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPageResponse'
  /api/v1/main_service/projects/hiring:
    get:
      tags:
        - Project
      summary: Поиск проектов, набирающих команду
      description: Возвращает проекты, у которых есть баннер поиска сотрудника. Можно фильтровать по специализации.
      operationId: searchProjectsHiring
      parameters:
        - name: specialization
          in: query
          description: Фильтр по специализации (частичное совпадение)
          schema:
            type: string
        - name: query
          in: query
          description: Поисковый запрос по названию проекта
          schema:
            type: string
        - name: page
          in: query
          required: true
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          required: true
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPageResponse'
  /api/v1/main_service/projects/looking-for:
    get:
      tags:
        - Project
      summary: Поиск проектов, ищущих ментора/инвестора/сооснователя
      description: Возвращает проекты, у которых есть баннер поиска указанного типа (mentor, investor, founder). Для mentor и founder можно фильтровать по области (domain).
      operationId: searchProjectsLookingFor
      parameters:
        - name: type
          in: query
          required: true
          description: "Тип поиска: mentor, investor, founder"
          schema:
            type: string
            enum:
              - mentor
              - investor
              - founder
        - name: domain
          in: query
          description: Фильтр по области (для mentor и founder)
          schema:
            type: string
        - name: query
          in: query
          description: Поисковый запрос по названию проекта
          schema:
            type: string
        - name: page
          in: query
          required: true
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          required: true
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPageResponse'
  /api/v1/main_service/me/projects:
    get:
      tags:
        - Project
      summary: Получение списка проектов прользователя
      description: Возвращает список проектов с поддержкой фильтрации по категории, поисковому запросу и статусу модерации. Поддерживает постраничный вывод.
      operationId: getMyProjects
      parameters:
        - name: role
          in: query
          description: Поиск проектов где участник имеет определенную роль
          schema:
            $ref: '#/components/schemas/MemberProjectRole'
        - name: query
          in: query
          description: Поисковый запрос по названию проекта
          schema:
            type: string
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          description: Размер страницы
          required: true
          schema:
            type: integer
            example: 10
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPageResponse'
  /api/v1/main_service/projects/{id}:
    get:
      tags:
        - Project
      summary: Получение проекта по ID
      description: Возвращает полную информацию о проекте по идентификатору.
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFullResponse'
    patch:
      tags:
        - Project
      summary: Обновление проекта
      description: Обновляет данные проекта по идентификатору.
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/ProjectRequestWithImage'
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFullResponse'
    delete:
      tags:
        - Project
      summary: Удаление проекта
      description: Удаляет проект по идентификатору.
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Проект успешно удален
  /api/v1/main_service/projects/{id}/likes:
    post:
      tags:
        - Project
      summary: Лайк/дизлайк проекта
      description: Переключает лайк текущего пользователя для указанного проекта.
      operationId: toggleProjectLike
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Состояние лайка после операции
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /api/v1/main_service/projects/{id}/likes/count:
    get:
      tags:
        - Project
      summary: Получение количества лайков проекта
      description: Возвращает текущее количество лайков проекта.
      operationId: getProjectLikesCount
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Количество лайков проекта
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /api/v1/main_service/projects/{id}/moderation-status:
    patch:
      tags:
        - Project
      summary: Обновление статуса модерации проекта
      description: Обновление статуса модерации проекта
      operationId: updateProjectModerationPassed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: boolean
              example: true
      responses:
        '200':
          description: Статус модерации успешно обновлен
        '404':
          description: Статья не найдена
...
components:
  schemas:
...
    ProjectPageResponse:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ProjectSummaryResponse'
...
    Page:
      type: object
      required:
        - content
        - number
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: Массив элементов текущей страницы
          items: { }
        number:
          type: integer
          example: 1
        size:
          type: integer
          example: 1
        totalElements:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1
    UserFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 42
        username:
          type: string
          description: Имя пользователя
          example: "Andrey"
    ProjectMemberFullResponse:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
          description: Идентификатор записи участника
          example: 7
        id:
          type: integer
          format: int64
          description: Идентификатор пользователя
          example: 15
        username:
          type: string
          description: Имя пользователя
          example: "User12345"
        role:
          type: string
          description: Роль пользователя в проекте
          example: "DevOps"
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор комментария
          example: 1001
        author:
          $ref: '#/components/schemas/UserFullResponse'
        text:
          type: string
          description: Текст комментария
          example: "User12345"
        createdAt:
          type: string
          description: Дата и время создания комментария
          example: "2026-01-24T14:32:10"
    ProjectMemberRequest:
      type: object
      required:
        - username
        - role
      properties:
        username:
          type: string
          description: Имя пользователя
          minLength: 1
          example: "User12345"
        role:
          type: string
          description: Роль пользователя в проекте
          minLength: 1
          example: "DevOps"
    ProjectFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор проекта
          example: 101
        owner:
          $ref: '#/components/schemas/UserFullResponse'
        members:
          type: array
          description: Участники проекта
          items:
            $ref: '#/components/schemas/ProjectMemberFullResponse'
        name:
          type: string
          description: Название проекта
          example: "Open Source Platform"
        description:
          type: string
          description: Описание проекта
          example: "Платформа для совместной разработки open-source проектов"
        logoUrl:
          type: string
          format: uri
          description: URL логотипа проекта
          example: "https://cdn.example.com/logos/project.png"
        category:
          type: string
          description: Категория проекта
          example: "Technology"
        likes:
          type: integer
          format: int64
          description: Количество лайков проекта
          example: 245
        hasSurvey:
          type: boolean
          description: Наличие опроса в проекте
          example: true
        comments:
          type: array
          description: Комментарии к проекту
          items:
            $ref: '#/components/schemas/CommentResponse'
        employeeSearches:
          type: array
          description: Поиск сотрудников
          items:
            $ref: '#/components/schemas/EmployeeSearchFullResponse'
        founderSearches:
          type: array
          description: Поиск сооснователей
          items:
            $ref: '#/components/schemas/FounderSearchFullResponse'
        investorSearches:
          type: array
          description: Поиск инвесторов
          items:
            $ref: '#/components/schemas/InvestorSearchFullResponse'
        mentorSearches:
          type: array
          description: Поиск менторов
          items:
            $ref: '#/components/schemas/MentorSearchFullResponse'
    ProjectSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор проекта
          example: 101
        owner:
          type: string
          description: Имя пользователя
          example: "User12345"
        members:
          type: array
          description: Участники проекта
          items:
            $ref: '#/components/schemas/ProjectMemberFullResponse'
        name:
          type: string
          description: Название проекта
          example: "Open Source Platform"
        logoUrl:
          type: string
          format: uri
          description: URL логотипа проекта
          example: "https://cdn.example.com/logos/project.png"
        category:
          type: string
          description: Категория проекта
          example: "Technology"
        likes:
          type: integer
          format: int64
          description: Количество лайков проекта
          example: 245
    ProjectRequest:
      type: object
      required:
        - name
        - description
        - categoryId
      properties:
        name:
          type: string
          description: Название проекта
          minLength: 1
          example: "Open Source Platform"
        description:
          type: string
          description: Описание проекта
          minLength: 1
          example: "Платформа для совместной разработки open-source проектов"
        categoryId:
          type: integer
          format: int64
          description: Идентификатор категории проекта
          example: 3
...
    EmployeeSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска сотрудника
          example: 1
        specialization:
          type: string
          description: Название специализации
          example: "Backend Developer"
        about:
          type: string
          description: Описание требований
          example: "Ищем опытного backend разработчика"
    FounderSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска сооснователя
          example: 1
        domain:
          type: string
          description: Направление деятельности
          example: "FinTech"
        about:
          type: string
          description: Описание требований
          example: "Ищем сооснователя с опытом в финансах"
    InvestorSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска инвестора
          example: 1
        about:
          type: string
          description: Описание требований
          example: "Ищем инвестора для seed-раунда"
    MentorSearchFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор записи поиска ментора
          example: 1
        domain:
          type: string
          description: Направление деятельности
          example: "EdTech"
        about:
          type: string
          description: Описание требований
          example: "Ищем ментора с опытом в EdTech"
    ImageRequest:
      type: string
      format: binary
    SeekingRole:
      type: string
      enum:
        - EMPLOYEE
        - MENTOR
        - INVESTOR
        - PARTNER
    MemberProjectRole:
      type: string
      enum:
        - OWNER
        - MEMBER
  requestBodies:
...
    ProjectRequestWithImage:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              project:
                $ref: '#/components/schemas/ProjectRequest'
              logo:
                $ref: '#/components/schemas/ImageRequest'
  parameters:
    XUserNameHeader:
      name: X-User-Name
      in: header
      description: "Имя пользователя, проставляется Gateway"
      required: true
      schema:
        type: string